# contract/openapi.yaml
...
openapi: 3.0.3
info:
  title: T-Shirt Store API
  version: 0.1.0
  description: |
    Contract for the T-Shirt Store API. Week 3 implements this document. If the
    implementation and the document disagree, the document is not out of date.

    Written against the schema in `4-database/3-erd/store.dbml`, which this branch
    carries since the ERD merge at `f96d62a`.

    **Money.** Every amount is an integer in minor units. 1999 means 19.99. The store
    is single-currency and prices are in USD. No response carries a currency field.

    **Tokens.** An access token is valid for 15 minutes. A refresh token is valid for
    7 days and rotates on every use. Sign-out deletes the presented device's refresh
    row. It does not revoke an access token the server already issued, so revocation
    lags by up to 15 minutes. Changing or resetting a password deletes every refresh row for
    that user.

    **Rate limits.** The server limits every operation by source address, in three tiers.
    Browsing carries the loosest limit. Sign-in carries a tighter limit. The three password
    operations carry the tightest limit. The server answers 429 with a `Retry-After` header
    when it refuses a request. This document declares the 429 response only where a client
    must handle it, but the server can answer 429 on any path.
servers:
  - url: http://localhost:3000/v1
    description: Local development. The /v1 prefix lives here, not on each path key.
...
security:
  - bearerAuth: []
paths:
...
  /users:
    post:
      tags: [auth]
      summary: Create an account
      description: |
        This operation is public. It does not need a token.

        Send an email address, a password and a name. The server creates a client
        account. The server returns the new account and a `Location` header that names
        it.

        The request does not accept a role. Every account this operation creates is a
        client account. This contract has no operation that changes a role, so a manager
        account is seeded.

        An email address that is already registered returns 409 with the `email-taken`
        problem type.

        The server limits the request rate. Too many requests return 429.
      operationId: createUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, firstName, lastName]
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 254
                  description: RFC 5321 4.5.3.1.3 caps a forward-path at 256 octets
                    including the punctuation, which leaves 254 for the address.
                  example: ana@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  maxLength: 128
                  description: The maximum is a cost bound, not a security rule. The
                    server hashes the value it receives. An unbounded field lets one
                    request consume a large amount of processor time.
                firstName:
                  type: string
                  maxLength: 100
                  example: Ana
                lastName:
                  type: string
                  maxLength: 100
                  example: Ramirez
      responses:
        '201':
          description: The server created the account.
          headers:
            Location:
              description: The URL of the new account.
              schema:
                type: string
              example: /v1/users/128
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/me/password:
    patch:
      tags: [auth]
      summary: Change the password of the signed-in account
      description: |
        Send the current password and the new password. The server changes the password.

        The server needs the current password. An access token is valid for 15 minutes,
        so a stolen token alone must not be enough to seize the account.

        A wrong current password returns 401. It is an authentication failure and not a
        permissions failure.

        On success the server deletes every refresh row for this user, including the row
        for the device that sent this request. The caller must sign in again. The server
        also sends an email to the account address.

        The server limits the request rate, because this operation accepts a password
        guess.
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                  format: password
                  minLength: 8
                  maxLength: 128
                newPassword:
                  type: string
                  format: password
                  minLength: 8
                  maxLength: 128
      responses:
        '204':
          description: The server changed the password and ended every session.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProblemType:
      type: string
      format: uri
      enum:
        # 401, and a client must tell these apart or it loops between refreshing and
        # sending the user back to the login screen.
        - https://tshirt.store/problems/invalid-credentials
        - https://tshirt.store/problems/access-token-expired
        - https://tshirt.store/problems/refresh-token-unknown
        # 409, and a client must tell these apart to show the right message.
        - https://tshirt.store/problems/email-taken
        - https://tshirt.store/problems/order-not-cancellable
        - https://tshirt.store/problems/insufficient-stock
    Problem:
      type: object
      # `type` is deliberately NOT required. The enum above names only the failures a
      # status code cannot tell apart, so a 404 or a 429 carries no member. RFC 9457
      # reads an absent `type` as `about:blank`.
      required: [title, status]
      properties:
        type:
          # `allOf` wraps the `$ref` because OpenAPI 3.0.3 ignores any sibling of a bare
          # `$ref`.
          allOf:
            - $ref: '#/components/schemas/ProblemType'
          description: >
            This member identifies the problem kind. Two requests that failed for the
            same reason carry the same value. An absent member means `about:blank`, so
            the status code is the whole story.
        title:
          type: string
          description: A short summary of the problem kind. It does not change between
            occurrences.
        status:
          type: integer
          description: The HTTP status code. It repeats the status line.
        detail:
          type: string
          description: An explanation of this specific occurrence.
        instance:
          type: string
          format: uri
          description: A URI reference that identifies this occurrence.
        errors:
          type: array
          description: An extension member. The server sends it only on a 400 from
            request validation. It carries one entry per rejected field.
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: must be a valid email address

    # ---- Sessions, DECISIONS.md items 8 and 10 ------------------------------
    # Returned by POST /auth/sessions and by POST /auth/refresh. The two return the same
    # shape because refresh replaces both tokens, so a client stores them the same way.
...
    User:
      type: object
      required: [id, email, firstName, lastName, role, createdAt]
      properties:
        id:
          type: integer
          example: 128
        email:
          type: string
          format: email
          example: ana@example.com
        firstName:
          type: string
          example: Ana
        lastName:
          type: string
          example: Ramirez
        role:
          $ref: '#/components/schemas/Role'
        createdAt:
          type: string
          format: date-time
          example: '2026-08-21T13:45:00Z'

    # ---- Catalog ------------------------------------------------------------
    # Two product shapes on purpose. The list entry carries a price summary and no
    # variants, so one page costs a fixed number of queries however many sizes a product
    # has. The detail carries the variants, and a client asks for it one product at a
    # time. That split is the answer to the N+1 reading.
...
    Role:
      type: string
      enum: [manager, client, delivery_person]

  # ---- Query parameters, DECISIONS.md item 3 ------------------------------
  # Every collection takes this pair. The bounds repeat PageMeta, so a caller reads the
  # same numbers on the request side and on the response side. A collection that does
  # not reference both returns an unbounded array.
...
  responses:
...
    BadRequest:
      description: >-
        The server cannot read the request. Either a body field failed validation, and
        then `errors` names each rejected field, or a path segment that must be an
        integer is not one. A path id that is a well formed integer but matches no row
        returns 404, not 400.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Validation failed
            status: 400
            detail: One or more fields did not pass validation.
            errors:
              - field: email
                message: must be a valid email address
    Unauthorized:
      description: No token, an expired token, or a refresh row that no longer exists.
      headers:
        WWW-Authenticate:
          schema:
            type: string
          description: Required on every 401.
          example: Bearer realm="tshirt-store"
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          # One named example per `ProblemType` member this status can carry. A client
          # that cannot separate these three loops between refreshing and sending the
          # user back to the login screen. DECISIONS.md item 2.
          examples:
            expired:
              summary: The access token expired
              value:
                type: https://tshirt.store/problems/access-token-expired
                title: Access token expired
                status: 401
                detail: Refresh the token and send the request again.
            credentials:
              summary: Wrong email or wrong password
              value:
                type: https://tshirt.store/problems/invalid-credentials
                title: Invalid credentials
                status: 401
                detail: The server did not accept this email and password.
            refresh:
              summary: The refresh token is unknown or already used
              value:
                type: https://tshirt.store/problems/refresh-token-unknown
                title: Refresh token unknown
                status: 401
                detail: Sign in again. The server ended every session for this user.
...
    Conflict:
      description: The request conflicts with the current state. The same request can
        succeed later.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          # The three 409 members. A client that cannot separate them shows the wrong
          # message. DECISIONS.md item 2.
          examples:
            emailTaken:
              summary: Sign-up against an existing account
              value:
                type: https://tshirt.store/problems/email-taken
                title: Email already registered
                status: 409
                detail: An account with this email already exists.
            insufficientStock:
              summary: Fewer units on hand than the request asks for
              value:
                type: https://tshirt.store/problems/insufficient-stock
                title: Not enough stock
                status: 409
                detail: This variant has 3 units on hand and the request asks for 5.
            notCancellable:
              summary: Cancel after the order shipped
              value:
                type: https://tshirt.store/problems/order-not-cancellable
                title: Order cannot be cancelled
                status: 409
                detail: This order already shipped.
...
    TooManyRequests:
      description: The caller reached the rate limit.
      headers:
        Retry-After:
          schema:
            type: string
          description: How long the caller must wait. RFC 9110 10.2.3 allows a number of
            seconds or an HTTP-date.
          example: '60'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Too many requests
            status: 429
            detail: Wait before you send this request again.
    InternalServerError:
      description: The server failed. The caller may retry the same request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Internal server error
            status: 500
            detail: The server failed to handle the request.